[Encode] Fix overflowed constant coverity issues
Fix overflowed constant coverity issues.

+++ media_driver/media_softlet/agnostic/Xe_M/Xe_XPM_base/codec/hal/enc/hevc/features/encode_hevc_vdenc_const_settings_xe_xpm_base.cpp
@@ -501,8 +501,8 @@ MOS_STATUS EncodeHevcVdencConstSettingsXe_Xpm_Base::SetVdencCmd2Settings()
                         return MOS_STATUS_SUCCESS;
                     }
 
-                    uint8_t tmp0 = 0;
-                    uint8_t tmp1 = 0;
+                    uint32_t tmp0 = 0;
+                    uint32_t tmp1 = 0;
 
                     if (m_hevcPicParams->CustomRoundingOffsetsParams.fields.EnableCustomRoudingIntra)
                     {
