+++ media_softlet/agnostic/common/codec/hal/enc/av1/features/encode_av1_rounding_table.cpp
@@ -129,8 +129,8 @@ MHW_SETPAR_DECL_SRC(VDENC_CMD2, Av1EncodeRoundingTable)
 #else
         params.extSettings.emplace_back(
             [av1BasicFeature](uint32_t *data) {
-                uint8_t tmp0 = av1BasicFeature->m_par65Intra & 0xf;
-                uint8_t tmp1 = av1BasicFeature->m_par65Inter & 0xf;
+                uint32_t tmp0 = av1BasicFeature->m_par65Intra & 0xf;
+                uint32_t tmp1 = av1BasicFeature->m_par65Inter & 0xf;
                 data[32] |= (tmp1 << 16);
                 data[32] |= (tmp1 << 20);
                 data[32] |= (tmp0 << 24);
