+++ media_softlet/agnostic/common/codec/hal/enc/hevc/features/encode_hevc_vdenc_const_settings.cpp
@@ -1874,8 +1874,8 @@ MOS_STATUS EncodeHevcVdencConstSettings::SetVdencLaCmd2Settings()
                         return MOS_STATUS_SUCCESS;
                     }
 
-                    uint8_t tmp0 = 0;
-                    uint8_t tmp1 = 0;
+                    uint32_t tmp0 = 0;
+                    uint32_t tmp1 = 0;
 
                     if (m_hevcPicParams->CustomRoundingOffsetsParams.fields.EnableCustomRoudingIntra)
                     {
